Add files via upload
Processing file.  This takes the serial data transmitted from Arduino and draws an ellipse to match the signal strength of the device.   Note that it is currently not complete. The final version would add a sonar sound that beeps according to the signal strength. Such that a strong signal will emit a loud beep, and a weak signal will emit a faint beep.  The Ellipse in its final form would be animated and move in real time according to the signal strength.